# Optimize LayeredSpatialMap.GetItemsAt