… TheSadRogue.Primitives/SpatialMaps/ReadOnlyLayeredSpatialMapItemsAtEnumerator.cs
… TheSadRogue.Primitives/SpatialMaps/SpatialMap.cs